+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Prettier-fix.
 - UserException to identify user runtime exceptions.
 - Immediate loginWithTrustedEmail for social login plugins.
+- AuthConstant::USER_ROUTE
 
 ### `Changed` for changes in existing functionality